requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pandas~=2.2.0
numpy>=1.23.5,<2.0.0
scikit-learn~=1.3.0
joblib==1.3.2
python-dotenv==1.0.1
alembic==1.13.1
python-multipart==0.0.9
tensorflow==2.15.0
jinja2>=3.0.0,<4.0.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
fastapi-mail==1.4.1
aiosmtplib==2.0.2
